+++ coeus-autograd/src/ops/nn/normalization/mod.rs
@@ -2,9 +2,6 @@ pub mod batchnorm;
 pub mod layernorm;
 pub mod rmsnorm;
 
-pub use batchnorm::{
-    batchnorm1d, batchnorm2d, batchnorm3d, BatchNorm1dArgs, BatchNorm1dNode, BatchNorm2dArgs,
-    BatchNorm2dNode, BatchNorm3dArgs, BatchNorm3dNode,
-};
+pub use batchnorm::{batchnorm1d, batchnorm2d, batchnorm3d, BatchNormArgs, BatchNormNode};
 pub use layernorm::{layernorm, LayerNormNode};
 pub use rmsnorm::{rmsnorm, RMSNormNode};