+++ coeus-nn/src/normalization/batchnorm1d.rs
@@ -196,7 +196,7 @@ impl<T: Float, B: coeus_ops::BackendOps<T> + Default> Module<T, B> for BatchNorm
             input,
             &self.weight,
             &self.bias,
-            coeus_autograd::BatchNorm1dArgs {
+            coeus_autograd::BatchNormArgs {
                 out_tensor,
                 x_hat,
                 xmu,
@@ -206,7 +206,7 @@ impl<T: Float, B: coeus_ops::BackendOps<T> + Default> Module<T, B> for BatchNorm
                 two_const: self.two_const.clone(),
                 n,
                 c,
-                l,
+                spatial: [l, 1, 1],
                 m,
             },
         )